Replace `DEB_BUILD_GNU_TYPE` with `DEB_HOST_GNU_TYPE`.
From the [CrossBuildPackagingGuidelines wiki page](https://wiki.debian.org/CrossBuildPackagingGuidelines) `DEB_BUILD_GNU_TYPE` is the architecture of the machine *building* the package, whereas `DEB_HOST_GNU_TYPE` is the architecture for which we are building the package.  I noticed this issue while cross-compiling an ARM container on my Mac.