Add alpine LTS builds for OTP 23 and 24
With alpine 3.13, the updated kernel headers break SCTP support
when the images are running on a host with < 5.5 kernel versions.
However the updates in alpine 3.13 have positive impact on memory
management and performance.
To keep SCTP working and benefit from the improvements on
alpine 3.13, this commit adds alpine-lts images that use 5.4 kernel
headers for building erlang OTP version 23 and 24.