make SCTP in images work older Linux kernels
Linux SCTP ioctls argument length checking is not forward compatible
with newer kernel headers.

The length of the SCTP ioctl argument is based on the size of some
SCTP structs. If code compiled with newer kernel headers, where the
SCTP structs are larger, is running on an older kernel and tries to
pass the a SCTP struct that is larger than what the running kernel
supports, then the ioctl results in an error.

Erlangs SCTP code does excatly that. That means that SCTP in a Erlang
VM that has been compiled with recent kernel headers will not work on
older kernels.

This change pulls in Linux 5.4.5 kernel headers (and only the kernel
headers) from Alpine 3.12. The resulting images will have working SCTP
support when running on Linux 5.4.5 and newer.